Separate files for each endpoint test cases
